fixed updated FIT

+++ src/Simulation.hh
@@ -18,7 +18,7 @@ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND 
 class Simulation {
 public:
 
-    Simulation( uint64_t interval_t, uint64_t scrub_interval_t, double fit_factor_t, uint test_mode_t, bool debug_mode_t, bool cont_running_t, uint64_t output_bucket_t, uint64_t turning_point_t);
+    Simulation( uint64_t interval_t, uint64_t scrub_interval_t, double fit_factor_t, uint test_mode_t, bool debug_mode_t, bool cont_running_t, uint64_t output_bucket_t, double turning_point_t);
 
 	void init( uint64_t max_s );
 	void reset( void );
@@ -39,7 +39,7 @@ protected:
 	bool debug_mode;
     bool cont_running;
     uint64_t m_output_bucket;
-    uint64_t turning_point;//xiao:turning point
+    double turning_point;//xiao:turning point
 
 	uint64_t stat_total_failures, stat_total_sims, stat_sim_seconds;
